pyproject.toml
[project]
name = "kube-boy"
version = "0.1.0"
description = ""
authors = [
    {name = "Frank Rosner",email = "3427394+FRosner@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "kubernetes (>=33.1.0,<34.0.0)",
    "openai (>=1.106.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
python-dotenv = "^1.1.1"

